+++ src/activitypub/valueflows/constants.rs
@@ -1,10 +1,9 @@
-// https://www.valueflo.ws/concepts/actions/#action-definitions
 pub const PURPOSE_OFFER: &str = "offer";
 
+// https://www.valueflo.ws/concepts/actions/#action-definitions
 pub const ACTION_DELIVER_SERVICE: &str = "deliverService";
 pub const ACTION_TRANSFER: &str = "transfer";
 
-pub const CLASS_CONTENT: &str = "https://www.wikidata.org/wiki/Q1260632";
 pub const CLASS_USER_GENERATED_CONTENT: &str = "https://www.wikidata.org/wiki/Q579716";
 
 // http://www.ontology-of-units-of-measure.org/resource/om-2/one